Find NetLogo Version

find_netlogo_version() attempts to locate the NetLogo version installed on the user's system.

It first tries to execute the NetLogo console with the --version argument to retrieve the version information. If the executable is not found, it attempts to extract the version number from the installation directory name.

Usage

find_netlogo_version(netlogo_home = find_netlogo_home())

Arguments

netlogo_home

(optional) A character string specifying the path to the NetLogo installation directory. If not provided, the function will try to find it automatically using find_netlogo_home(). (default: find_netlogo_home()).

Value

A character string specifying the NetLogo version installed on the user's system. If the version cannot be determined, an empty string is returned.

See also

Examples

if (FALSE) { # \dontrun{
  find_netlogo_version()
} # }